# Conflicts:
#	src/main/java/uz/pdp/olx/controller/CategoryController.java
#	src/main/java/uz/pdp/olx/exception/CategoryNotFoundException.java
#	src/main/java/uz/pdp/olx/exception/handler/GlobalExceptionHandler.java
#	src/main/java/uz/pdp/olx/repository/CategoryRepository.java
#	src/main/java/uz/pdp/olx/repository/ProductRepository.java
#	src/main/java/uz/pdp/olx/service/ImageService.java
#	src/main/java/uz/pdp/olx/service/ProductService.java

+++ src/main/java/uz/pdp/olx/service/ProductService.java
@@ -2,16 +2,14 @@ package uz.pdp.olx.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import uz.pdp.olx.dto.ProductDto;
 import uz.pdp.olx.dto.ProductSaveDto;
 import uz.pdp.olx.dto.ProductUpdateDto;
 import uz.pdp.olx.dto.UserDto;
 import uz.pdp.olx.enitiy.Product;
-import uz.pdp.olx.exception.CategoryNotFoundException;
-import uz.pdp.olx.exception.ProductNotFoundException;
-import uz.pdp.olx.exception.UserNotFoundException;
+import uz.pdp.olx.exception.NotFoundException;
+import uz.pdp.olx.exception.NullOrEmptyException;
 import uz.pdp.olx.repository.CategoryRepository;
 import uz.pdp.olx.repository.ProductRepository;
 import uz.pdp.olx.repository.UserRepository;
@@ -37,8 +35,8 @@ public class ProductService {
         product.setItemCondition(productSaveDto.getItemCondition());
         product.setPrice(productSaveDto.getPrice());
         product.setItemCondition(productSaveDto.getItemCondition());
-        product.setUser(userRepository.findById(productSaveDto.getUserId()).orElseThrow(()-> UserNotFoundException.byId(productSaveDto.getUserId())));
-        product.setCategory(categoryRepository.findById(productSaveDto.getCategoryId()).orElseThrow(CategoryNotFoundException::new));
+        product.setUser(userRepository.findById(productSaveDto.getUserId()).orElseThrow(()-> new NotFoundException("User")));
+        product.setCategory(categoryRepository.findById(productSaveDto.getCategoryId()).orElseThrow(() -> new NotFoundException("Category")));
         productRepository.save(product);
         return new ProductDto(
                 product.getId(),
@@ -54,36 +52,58 @@ public class ProductService {
 
 
 public Optional<ProductDto> findByProductId(Long id){
-        return Optional.ofNullable(productRepository.finBYdId(id).orElseThrow(ProductNotFoundException::new));
+        if (id == null) {
+            throw new NullOrEmptyException("Id");
+        }
+        return Optional.ofNullable(productRepository.finBYdId(id).orElseThrow(
+                () -> new NotFoundException("Product")
+        ));
 
 }
 
 
 public Optional<List<ProductDto>>findAllByProductTitle(String title){
+    if (title == null) {
+        throw new NullOrEmptyException("Title");
+    }
         return Optional.ofNullable(productRepository.findAllByTitle(title).orElseThrow(
-                ProductNotFoundException::new
+                () -> new NotFoundException("Product")
         ));
 }
 public Boolean deleteByProductId(Long id){
+        if (id == null) {
+            throw new NullOrEmptyException("Id");
+        }
         if(productRepository.existsById(id)){
              productRepository.deleteById(id);
              return true;
         }
         else {
-            throw new ProductNotFoundException();
+            throw new NotFoundException("Product");
         }
 }
 public ProductDto updateProduct(ProductUpdateDto productUpdateDto){
+        if (productUpdateDto.getTitle() == null) {
+            throw new NullOrEmptyException("Product Title");
+        }
+        if (productUpdateDto.getDescription() == null) {
+            throw new NullOrEmptyException("Product Description");
+        }
+        if (productUpdateDto.getPrice() == null) {
+            throw new NullOrEmptyException("Product Price");
+        }
+        if (productUpdateDto.getCategoryId() == null) {
+            throw new NullOrEmptyException("Product CategoryId");
+        }
     Product product = productRepository.findById(productUpdateDto.getId())
-            .orElseThrow(ProductNotFoundException::new);
+            .orElseThrow(() -> new NotFoundException("Product"));
     product.setTitle(productUpdateDto.getTitle());
     product.setDescription(productUpdateDto.getDescription());
     product.setIsActive(productUpdateDto.getIsActive());
     product.setPrice(productUpdateDto.getPrice());
     product.setItemCondition(productUpdateDto.getItemCondition());
     product.setCategory(categoryRepository.findById(productUpdateDto.getCategoryId())
-            .orElseThrow(CategoryNotFoundException::new));
-    productRepository.save(product);
+            .orElseThrow(() -> new NotFoundException("Category")));
     return new ProductDto(product);
 }
